[B] Refactor Scan file to be in a folder of itself
