+++ src/gui/gui_scan_page/gui_scan_manager.py
